feat: melhoria PetsController com validação de admin e uso de DTO

+++ src/pets/controllers/pets.controller.ts
@@ -7,7 +7,10 @@ import {
   Post,
   Put,
   Query,
+  UnauthorizedException,
+  UploadedFile,
   UseGuards,
+  UseInterceptors,
 } from '@nestjs/common';
 import { PetsService } from '../services/pets.service';
 import { Pet } from '../entities/pet.entity';
@@ -16,6 +19,9 @@ import { CurrentUser } from 'src/auth/decorators/current-user.decorator';
 import { JwtAuthGuard } from 'src/auth/guards/jwt-auth.guard';
 import { BlackbazeService } from '../services/blackbaze.service';
 import { FilterAdocaoDto } from '../dtos/FilterAdocaoDto';
+import { IsAdmin } from 'src/auth/decorators/is-admin.decorator';
+import { UpdatePetDto } from '../dtos/UpdatePetDto';
+import { FileInterceptor } from '@nestjs/platform-express';
 
 @Controller('pets')
 export class PetsController {
@@ -26,15 +32,8 @@ export class PetsController {
 
   @UseGuards(JwtAuthGuard)
   @Get()
-  findAll(
-    @CurrentUser('ongId') ongId: string,
-    @Query('page') page?: string,
-    @Query('limit') limit?: string,
-  ) {
-    return this.petsService.findAll(ongId, {
-      page: page ? parseInt(page, 10) : 1,
-      limit: limit ? parseInt(limit, 10) : 10,
-    });
+  findAll(@CurrentUser('ongId') ongId: string) {
+    return this.petsService.findAll(ongId);
   }
 
   @Get('adocao')
@@ -42,16 +41,16 @@ export class PetsController {
     return this.petsService.findAllAdocao(query);
   }
 
-  // @Get('adocao')
-  // findAllAdocao() {
-  //   return this.petsService.findAllAdocao();
-  // }
-
   @Get('upload-url')
   async getUploadUrl() {
     const res = await this.blackbazeService.getUploadUrl();
     console.log(res);
     return res;
+  }
+
+  @Get('download-img-auth')
+  async getDownloadAuth() {
+    return this.blackbazeService.getDownloadAuthorization();
   }
 
   @UseGuards(JwtAuthGuard)
@@ -68,7 +67,14 @@ export class PetsController {
 
   @UseGuards(JwtAuthGuard)
   @Delete(':id')
-  remove(@Param('id') id: string, @CurrentUser('ongId') ongId: string) {
+  remove(
+    @Param('id') id: string,
+    @CurrentUser('ongId') ongId: string,
+    @IsAdmin('is_admin') isAdmin: boolean,
+  ) {
+    if (!isAdmin) {
+      throw new UnauthorizedException('Ação restrita para administradores');
+    }
     return this.petsService.remove(id, ongId);
   }
 
@@ -76,7 +82,7 @@ export class PetsController {
   @Put(':id')
   update(
     @Param('id') id: string,
-    @Body() pet: Pet,
+    @Body() pet: UpdatePetDto,
     @CurrentUser('ongId') ongId: string,
   ) {
     return this.petsService.update(id, pet, ongId);
